+++ app/charts/column/overlay-columns.tsx
@@ -1,44 +1,61 @@
+import { useCallback } from "react";
+
 import { ColumnsState } from "@/charts/column/columns-state";
 import { ComboLineColumnState } from "@/charts/combo/combo-line-column-state";
 import { useChartState } from "@/charts/shared/chart-state";
 import { useInteraction } from "@/charts/shared/use-interaction";
 import { Observation } from "@/domain/data";
+import useEvent from "@/utils/use-event";
 
-export const InteractionColumns = () => {
+export const InteractionColumns = ({ temporal }: { temporal?: boolean }) => {
   const [, dispatch] = useInteraction();
-
-  const { chartData, bounds, getX, xScaleInteraction } = useChartState() as
-    | ColumnsState
-    | ComboLineColumnState;
-  const { margins, chartHeight } = bounds;
-
-  const showTooltip = (d: Observation) => {
+  const {
+    chartData,
+    bounds: { margins, chartHeight },
+    getX,
+    getXAsDate,
+    formatXDate,
+    xScaleInteraction,
+  } = useChartState() as ColumnsState | ComboLineColumnState;
+  const showTooltip = useEvent((d: Observation) => {
     dispatch({
       type: "INTERACTION_UPDATE",
       value: { interaction: { visible: true, d } },
     });
-  };
-  const hideTooltip = () => {
+  });
+  const hideTooltip = useEvent(() => {
     dispatch({
       type: "INTERACTION_HIDE",
     });
-  };
+  });
+  const getXValue = useCallback(
+    (d: Observation) => {
+      return temporal ? formatXDate(getXAsDate(d)) : getX(d);
+    },
+    [formatXDate, getX, getXAsDate, temporal]
+  );
+
   return (
     <g transform={`translate(${margins.left} ${margins.top})`}>
-      {chartData.map((d, i) => (
-        <rect
-          key={i}
-          x={xScaleInteraction(getX(d)) as number}
-          y={0}
-          width={xScaleInteraction.bandwidth()}
-          height={Math.max(0, chartHeight)}
-          fill="hotpink"
-          fillOpacity={0}
-          stroke="none"
-          onMouseOut={hideTooltip}
-          onMouseOver={() => showTooltip(d)}
-        />
-      ))}
+      {chartData.map((d, i) => {
+        const x = getXValue(d);
+        const xScaled = xScaleInteraction(x) as number;
+
+        return (
+          <rect
+            key={i}
+            x={xScaled}
+            y={0}
+            width={xScaleInteraction.bandwidth()}
+            height={Math.max(0, chartHeight)}
+            fill="hotpink"
+            fillOpacity={0}
+            stroke="none"
+            onMouseOut={hideTooltip}
+            onMouseOver={() => showTooltip(d)}
+          />
+        );
+      })}
     </g>
   );
 };