+++ CHANGELOG.md
@@ -13,6 +13,11 @@ You can also check the
 
 - Fixes
   - Column line combo charts now work correctly with temporal entity X dimension
+  - Default min custom scale domain value now defaults to 0 if no negative
+    values are present
+  - Chart area no longer overflows below X or Y axes when setting custom scale
+    domains
+  - Band axis ticks should now be aligned correctly
 
 # 5.7.6 - 2025-05-06